requirements.txt
streamlit
google-generativeai
srt